Temporarily remove visibility icons and update spans' class names
Update spans to include class namnes. This is so we don't have any
text content within the HTML span elements. The spans' ::before and
::after content will be strictly dictated by CSS.

Also reordered the future visibility icon span to be after the input.
Since I reordered the span, I had to update the advanced sibiling
selectors for the pre-existing "valid" and "invalid" icons.

Remove associated visibility toggle in JS.